@@ -30,7 +30,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  stream_chat_flutter: ^0.2.21 
+  stream_chat_flutter: ^6.8.0 
 
 
   # The following adds the Cupertino Icons font to your application.
